feat: Header, 프로필 기능 수정
